Added queue specific configs for handlers.
This will allow us to do thing like have the AtomPub handler
route events from a specific queue to a different atompub endpoint.
